+++ src/ofxGamuza/src/gamuza/gamuzaAudio.h
@@ -11,7 +11,7 @@ int						inputBufferCounter;
 
 //--------------------------------------------------------------
 void gamuzaMain::setupAudio(){
-	
+
 	//////////////////////////////////////////
 	// inicializo el DSP de audio
 	computeAudioInput		= false;
@@ -26,31 +26,31 @@ void gamuzaMain::setupAudio(){
     recSize                 = 0;
     recChannel              = 0;
     recordingInput          = false;
-	
+
 	printf("\n");
 	soundStream.setDeviceID(audioDevID);
 	soundStream.setup(this,audioOutputChannels,audioInputChannels,audioSamplingRate,audioBufferSize, audioNumBuffers);
-	
+
 	inputAudioCH = new audioInputChannel[audioInputChannels];
-	
+
 	for(int i=0; i<audioInputChannels; i++){
 		inputAudioCH[i].setupChannel(i,audioSamplingRate,audioBufferSize,audioInputChannels,fftWindowUse);
 	}
-	
+
 	soundStream.stop();
-	
+
 	if(audioOutputChannels > 0){
 		gamuzaDSP.setupDSP(audioOutputChannels);
 		gamuzaAMP.setup(audioOutputChannels);
         exportAudio             = true;
 	}
 	//////////////////////////////////////////
-	
+
 }
 
 //--------------------------------------------------------------
 void gamuzaMain::audioIn(float * input, int bufferSize, int nChannels){
-	
+
 	// samples are "interleaved"
 	//  L	R	L	R	L	R
 	// CH1 CH2 CH1 CH2 CH1 CH2
@@ -62,26 +62,26 @@ void gamuzaMain::audioIn(float * input, int bufferSize, int nChannels){
 			for(int i = 0; i < audioInputChannels; i++){
 				if(inputAudioCH[i].captureAudio && inputAudioCH[i].computeChannel){
 					inputAudioCH[i].captureChannel(input);
-                    
+
                     // input recording (MONO CHANNEL)
                     if(recordingInput && i == recChannel){
                         for (int j = 0; j < bufferSize; j++){
                             recBuffer.push_back(input[j*nChannels + recChannel]);
                         }
                     }
-                    
+
 				}
 			}
 			///////////////////////////////////////////
 			inputBufferCounter++;
 		}
 	}
-	
+
 }
 
 //--------------------------------------------------------------
 void gamuzaMain::audioOut(float * output, int bufferSize, int nChannels){
-	
+
 	// samples are "interleaved"
 	//  L	R	L	R	L	R
 	// CH1 CH2 CH1 CH2 CH1 CH2
@@ -90,45 +90,45 @@ void gamuzaMain::audioOut(float * output, int bufferSize, int nChannels){
 		if(audioOutputChannels > 0){
 			gamuzaDSP.clearBuffer(output, bufferSize);
 			///////////////////////////////////////////
-            
+
 			// audio synthesis
-			for(int i = 0; i < audioModules.size(); i++){
+			for(int i = 0; i < (int)audioModules.size(); i++){
 				audioModules[i].addToSoundBuffer(output, bufferSize, gamuzaDSP.numOscInCh);
 			}
-            
+
             // audio samples
-            for(int i = 0; i < audioSamples.size(); i++){
+            for(int i = 0; i < (int)audioSamples.size(); i++){
                 if(audioSamples[i].getIsPlaying()){
                     audioSamples[i].addToSoundBuffer(output, bufferSize);
                 }
             }
-            
+
             // audio input recordings
-            for(int i = 0; i < inputRecSamples.size(); i++){
+            for(int i = 0; i < (int)inputRecSamples.size(); i++){
                 if(inputRecSamples[i].getIsPlaying()){
                     inputRecSamples[i].addToSoundBuffer(output, bufferSize);
                 }
             }
-            
+
             // audio export
             if(isExporting){
                 audioExport.addSamples(output,bufferSize*nChannels);
             }
-            
+
 			///////////////////////////////////////////
 			gamuzaAMP.addToSoundBuffer(output, bufferSize, mainVolume);
-            
+
             ///////////////////////////////////////////
             // visualizing output signal
 			if(ofGetFrameNum()%2 == 0){
 				memcpy(outputBufferCopy, output, bufferSize * nChannels * sizeof(float));
 			}
             ///////////////////////////////////////////
-			
+
 			outputBufferCounter++;
 		}
 	}
-	
+
 }
 
 //--------------------------------------------------------------
@@ -185,13 +185,13 @@ void gamuzaMain::resetAudioOutput(){
     if(recBuffer.size() > 0){
 		recBuffer.clear();
     }
-    
+
     recSize         = 0;
-    
+
 	audioModulesNum = 0;
     audioSamplesNum = 0;
     inputRecordingsNum = 0;
-    
+
 	gamuzaDSP.resetOsc();
 }
